requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg[binary]
python-dotenv
google-genai
